Created function to let our site know extension details
When called in the extension this should create an element on our site
(bufferapp.com/app) that has the extension version information. If the
user does not have this we can conclude that they either don't have it
installed or don't have it up to date.

With the version information being passed through as well we can
potentially give people notes to updgrade their plugin for a better
experience, especially the Firefox and Safari plugins that don't appear
to be on the store anywhere.