Buffer button does not seem to work when triggering it from the "Pin It" extension
Hey @pioul! Would love your :eyes: and thoughts on this one :smile: It
looks like it might fix #buffer-web/8353

Sometimes, the $img element is not being found traversing the tree as we
are doing. The only fix I could think of was to default to grab the
$('.pinImg') directly on cases it was undefined with the other way. Not
sure if it fixes things, and my extension-fu is a bit rusty anyway haha
:sweat_smile:

Thanks so much for taking a look!